# Reduce sleep between reading server logs to improve performance for heavy-logging Laravel application using a stderr/stdout logging driver.