--- a/src/Commands/Concerns/InteractsWithServers.php
+++ b/src/Commands/Concerns/InteractsWithServers.php
@@ -28,6 +28,8 @@ trait InteractsWithServers
         $watcher = $this->startServerWatcher();
 
         try {
+            $usleepBetweenIterations = config('octane.usleep_between_writing_server_output') ?? $_ENV['LARAVEL_OCTANE_USLEEP_BETWEEN_WRITING_SERVER_OUTPUT'] ?? 10 * 1000;
+
             while ($server->isRunning()) {
                 $this->writeServerOutput($server);
 
@@ -45,7 +47,7 @@ trait InteractsWithServers
                     return 1;
                 }
 
-                usleep(500 * 1000);
+                usleep($usleepBetweenIterations);
             }
 
             $this->writeServerOutput($server);